requirements.txt
Django==4.2.7
black==24.10.0
django-environ==0.11.2
psycopg2-binary==2.9.10
mysqlclient==2.2.7
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
dj-rest-auth==7.0.0
django-allauth[socialaccount]==65.2.0
djangorestframework-simplejwt==5.3.1